+++ src/main/java/com/ruoyi/project/count/work/service/IWorkService.java
@@ -1,7 +1,8 @@
 package com.ruoyi.project.count.work.service;
 
-import java.util.List;
 import com.ruoyi.project.count.work.domain.Work;
+
+import java.util.List;
 
 /**
  * 教学工作量Service接口
@@ -17,7 +18,7 @@ public interface IWorkService
      * @param id 教学工作量ID
      * @return 教学工作量
      */
-    public Work selectWorkById(Long id);
+    public Work selectWorkById(Integer id);
 
     /**
      * 查询教学工作量列表
@@ -57,5 +58,5 @@ public interface IWorkService
      * @param id 教学工作量ID
      * @return 结果
      */
-    public int deleteWorkById(Long id);
+    public int deleteWorkById(Integer id);
 }
